@@ -5,56 +5,69 @@ import org.json.JSONObject;
 
 import java.util.regex.Pattern;
 
-public class Scanner {
+/**
+ * Authenticates a ticket code with the server.
+ * Based on the code from the Scanner class from the original Crossoff project.
+ */
+public class TicketAuthenticator implements HttpClient.ReplyHandler {
 
-    // if running as Android Virtual Device, 10.0.2.2 is the magic address of host system localhost
-    public static final String CROSSOFF_SERVER_HOST_ADDR = "10.0.2.2";
-    public static final int CROSSOFF_SERVER_PORT = 8080;
+    ResultHandler consumer;
+    // acceptable barcodes: subset of Code 128 printable characters (no / + or [space])
+    private final static Pattern VALID_TICKET_CODE = Pattern.compile("^[A-Z0-9\\-\\$\\%\\*]+$");
 
-    // acceptable barcodes: subset of Code 39 printable characters (no / + or [space])
-    private static Pattern VALID_TICKET_CODE = Pattern.compile("^[A-Z0-9\\-\\$\\%\\*]+$");
+    public TicketAuthenticator(ResultHandler consumer){
+        this.consumer = consumer;
+    }
 
-    public static boolean validTicketCode(String code) {
+    public static boolean validCodeFormat(String code) {
         return code != null && VALID_TICKET_CODE.matcher(code).matches();
     }
 
-    public static void scanTicket(String ticketCode, ResultHandler consumer) {
-        if (!validTicketCode(ticketCode)) {
-            consumer.accept("Weird barcode scanned. "
+    /*
+     * Validates a ticket code against a specified format and with the server at the given URL.
+     * [baseUrl] is the URL of the server to connect to for validation.
+     * [ticketCode] is the ticket code to validate.
+     */
+    public void validateTicketCode(String baseUrl, String ticketCode) {
+        if (!validCodeFormat(ticketCode)) {
+            if(consumer != null)
+                consumer.postServerResult("Weird barcode scanned. "
                     + "Make sure camera is pointed at ticket!\n(scan = " + ticketCode + ")");
         } else {
-            serverScan(ticketCode, consumer);
+            serverValidation(baseUrl, ticketCode);
         }
     }
 
-    private static void serverScan(String ticketCode, final ResultHandler consumer) {
+    private void serverValidation(String baseUrl, String ticketCode) {
         HttpClient httpClient = new HttpClient();
-        String url = "http://" + CROSSOFF_SERVER_HOST_ADDR + ":" + CROSSOFF_SERVER_PORT + "/tickets/" + ticketCode;
-        httpClient.post(url, "", new HttpClient.Handler() {
-            @Override
-            public void accept(int statusCode, String body) {
-                boolean accepted = false;
-                try {
-                    String scanResult;
-                    if (body == null) {
-                        scanResult = "ERROR: Communication with server failed. Check network and try again. ("+statusCode+") " + url;
-                    } else if (statusCode == 200) {
-                        scanResult = parseScanResult(body);
-                    } else {
-                        scanResult = "ERROR: Unexpected response from server: HTTP " + statusCode;
-                    }
-                    accepted = true;
-                    consumer.accept(scanResult);
-                } finally {
-                    if (!accepted) {
-                        consumer.accept("Internal error!");
-                    }
-                }
-            }
-        });
+        String url = baseUrl + "/tickets/" + ticketCode;
+        httpClient.post(url, "", this);
     }
 
-    private static String parseScanResult(String body) {
+    // HttpClient.ReplyHandler implementation
+    @Override
+    public void serverReply(int statusCode, String body) {
+        boolean accepted = false;
+        try {
+            String serverResult;
+            if (body == null) {
+                serverResult = "ERROR: Communication with server failed. Check network and try again. ("+statusCode+")";
+            } else if (statusCode == 200) {
+                serverResult = parseServerResult(body);
+            } else {
+                serverResult = "ERROR: Unexpected response from server: HTTP " + statusCode;
+            }
+            accepted = true;
+            if(consumer != null)
+                consumer.postServerResult(serverResult);
+        } finally {
+            if (!accepted && consumer != null) {
+                consumer.postServerResult("Internal error!");
+            }
+        }
+    }
+
+    public String parseServerResult(String body) {
         String result;
         JSONObject scanResult = null;
         try {
@@ -105,8 +118,8 @@ public class Scanner {
         return result;
     }
 
+    //Handler for sending the server result to the consumer
     public interface ResultHandler {
-
-        void accept(final String scanResult);
+        void postServerResult(final String serverResult);
     }
 }
